% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasdetection-package.R
\docType{package}
\name{biasdetection-package}
\alias{biasdetection}
\alias{biasdetection-package}
\title{biasdetection: Bias Detection for Stated Preference Experiments}
\description{
Identifies the different types of biases in a stated preference dataset.
}
\author{
\strong{Maintainer}: Rupanjan Chakraborty \email{rupanjan.chakraborty@tum.de}

}
\keyword{internal}
