% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasdetection_functions.R
\name{create_plots_ers}
\alias{create_plots_ers}
\title{Create Plots Extreme Response Style}
\usage{
create_plots_ers(ers_data, likert_columns)
}
\arguments{
\item{ers_data}{Resulting data from the ers_function}

\item{likert_columns}{A list containing the column names to use for ERS
create_plots_ers()}
}
\description{
This function creates plots from the resulting data in extreme response style
(question wise bar plot and respondent wise bar plot (for positive, negative and total),
kernel density plot for respondent wise).
}
