% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasdetection_functions.R
\name{ers_function}
\alias{ers_function}
\title{Extreme Responding Style}
\usage{
ers_function(csv_file, likert_columns, max_value, min_value)
}
\arguments{
\item{csv_file}{The file we want to process}

\item{likert_columns}{A list containing the column names to use for ERS}

\item{max_value}{Maximum rate in the likert scale}

\item{min_value}{Minimum rate in the likert scale}
}
\value{
A list containing the updated data
ers_function()
}
\description{
Extreme Responding Style
}
