% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasdetection_functions.R
\name{get_time_groups}
\alias{get_time_groups}
\title{Get number of time groups}
\usage{
get_time_groups(csv_file)
}
\arguments{
\item{csv_file}{The .csv file we want to process}
}
\value{
Number of time groups
get_time_groups()
}
\description{
This function reads the .csv file which is originated from Lime Survey
and calculates the number of time groups automatically for detecting random responses based on the time spent in filling out parts of the survey.
}
