% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasdetection_functions.R
\name{lrs_general_function}
\alias{lrs_general_function}
\title{Lexicographic Response Style Function}
\usage{
lrs_general_function(
  csv_file,
  total_scenarios,
  num_blocks,
  scenarios,
  alternatives,
  attribute,
  attribute_short
)
}
\arguments{
\item{csv_file}{The file we want to process}

\item{total_scenarios}{A list containing all the Stated Preference scenarios}

\item{num_blocks}{The amount of blocks on which the scenarios will be distributed}

\item{scenarios}{A list containing the relevant Stated Preference scenarios for the specific attribute of Lexicographic Response Styles}

\item{alternatives}{A list containing the alternatives for the Stated Preference experiment}

\item{attribute}{The attribute chosen for Lexicographic Response Styles}

\item{attribute_short}{The short form of the attribute name (for table names)
lrs_general_function()}
}
\description{
This function detects the responses employing lexicographic response styles, where the respondents choose the option for stated preference
scenarios based on the best attribute level.
}
