% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasdetection_functions.R
\docType{data}
\name{mid_value}
\alias{mid_value}
\title{Mid point Response Styles}
\format{An object of class \code{numeric} of length 1.}
\usage{
mid_value
}
\arguments{
\item{csv_file}{The file we want to process}

\item{likert_columns}{A list containing the column names to use for ERS}

\item{mid_value}{Mid rate in the likert scale}
}
\value{
A list containing the updated data
mrs_function()
}
\description{
Mid point Response Styles
}
\keyword{datasets}
