% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasdetection_functions.R
\name{mrs_function}
\alias{mrs_function}
\title{Mid point Response Styles}
\usage{
mrs_function(csv_file, likert_columns, mid_value)
}
\arguments{
\item{csv_file}{The file we want to process}

\item{likert_columns}{A list containing the column names to use for Midpoint Response Styles}

\item{mid_value}{Mid rate in the likert scale}
}
\value{
A list containing the updated data
mrs_function()
}
\description{
This function detects the midpoint response styles in responses for the Likert scale questions.
}
\examples{
\dontrun{
mrs_function(file.csv, list("ParkingCosts.SQ001.",
                            "CongestionCosts.SQ001.",
                            "LowIncLikert.SQ001.",
                            "AffordableTrLikert.SQ001."), 3)
}
}
