% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasdetection_functions.R
\name{random_responding_time_filter}
\alias{random_responding_time_filter}
\title{Random responding time filter}
\usage{
random_responding_time_filter(rr_data, excluded_time_groups, num_time_groups)
}
\arguments{
\item{rr_data}{A list containing unique responses (based on id) from the original panel data}

\item{excluded_time_groups}{A list containing the indices of excluded time groups}

\item{num_time_groups}{Number of time groups}
}
\value{
A list containing the updated data
random_responding_time_filter()
}
\description{
For each Time Group and for each respondent, the results of the random responses are recorded as binary.
It also creates 2 additional columns (RR_Total and percent_RR_Total) containing the sum of the random responses for each
Time Group, and the total percentage of random responses for each respondent, respectively.
}
