% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasdetection_functions.R
\name{read_csv}
\alias{read_csv}
\title{Read .csv file}
\usage{
read_csv(csv_file)
}
\arguments{
\item{csv_file}{The file we want to process}
}
\value{
A list containing the data from the .csv file
read_csv()
}
\description{
This function reads the .csv file which is originated from Lime Survey.
It also calculates the number of time groups automatically for detecting random responses based on the time spent in filling out parts of the survey.
}
