% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasdetection_functions.R
\name{rr_function}
\alias{rr_function}
\title{Write the updated .csv file}
\usage{
rr_function(csv_file, excluded_time_groups, num_time_groups)
}
\arguments{
\item{csv_file}{The file we want to process}

\item{excluded_time_groups}{A list containing the indices of excluded time groups}

\item{num_time_groups}{Number of time groups}
}
\value{
A list containing the updated data
rr_function()
}
\description{
Given a .csv file with panel data, excludes duplicate entries, applies time filter for random responses
and writes the result in another .csv file.
}
\examples{
\dontrun{
rr_function(file.csv, list(2,4,5), 10)
}
}
