% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasdetection_functions.R
\name{attribute_non_attendance_function}
\alias{attribute_non_attendance_function}
\title{Attribute Non Attendance Function (for Inconsistent Bias and Non Trading)}
\usage{
attribute_non_attendance_function(
  csv_file,
  total_scenarios,
  num_blocks,
  scenarios,
  alternatives,
  attribute,
  attribute_short
)
}
\arguments{
\item{csv_file}{The file we want to process}

\item{total_scenarios}{A list containing all the SP scenarios}

\item{num_blocks}{The amount of blocks on which the scenarios will be distributed}

\item{scenarios}{A list containing the relevant SP scenarios for the specific attribute of LRS}

\item{alternatives}{A list containing the alternatives for the SP experiment}

\item{attribute}{The attribute chosen for LRS}

\item{attribute_short}{The short form of the attribute name (for table names)
attribute_non_attendance_function()}
}
\description{
Attribute Non Attendance Function (for Inconsistent Bias and Non Trading)
}
\examples{
\dontrun{
attribute_non_attendance_function(file.csv, list(list("S1", "S2"),
                                                 list("S3", "S4"),
                                                 list("S5", "S6"),
                                                 list("S7", "S8")),
                                                 4,
                                                 list(list("S1"),
                                                      list("S4"),
                                                      list("S6"),
                                                      list("S7")),
                                                      "Cheapest_Cost",
                                                      "CC")
attribute_non_attendance_function(file.csv, list(list("S1", "S2", "S3", "S4"),
                                                 list("S5", "S6", "S7", "S8")),
                                                 2,
                                                 list(list("S1", "S3"),
                                                      list("S7", "S8")),
                                                      "Fastest_Travel_Time",
                                                      "FTT")
}
}
