% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasdetection_functions.R
\name{create_plots_rr}
\alias{create_plots_rr}
\title{Create Plots Random Responding}
\usage{
create_plots_rr(rr_data, excluded_time_groups, num_time_groups)
}
\arguments{
\item{rr_data}{A list containing unique responses (based on id) from the original panel data}

\item{excluded_time_groups}{A list containing the indices of excluded time groups}

\item{num_time_groups}{Number of time groups
create_plots_rr()}
}
\description{
This function creates plots from the resulting data in random responding
(bar plot for group wise and bar plot, density plot and histogram for respondent wise)
}
