% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasdetection_functions.R
\name{lrs_general_function}
\alias{lrs_general_function}
\title{Lexicographic Response Style Function}
\usage{
lrs_general_function(
  csv_file,
  total_scenarios,
  scenarios,
  alternatives,
  attribute,
  attribute_short
)
}
\arguments{
\item{csv_file}{The file we want to process}

\item{total_scenarios}{A list containing all the SP scenarios}

\item{scenarios}{A list containing the relevant SP scenarios for the specific attribute of LRS}

\item{alternatives}{A list containing the alternatives for the SP experiment}

\item{attribute}{The attribute chosen for LRS}

\item{attribute_short}{The short form of the attribute name (for table names)}
}
\value{
A list containing the updated data
lrs_general_function()
}
\description{
Lexicographic Response Style Function
}
