% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasdetection_functions.R
\name{lrs_general_function}
\alias{lrs_general_function}
\title{Lexicographic Response Style Function}
\usage{
lrs_general_function(
  csv_file,
  total_scenarios,
  num_blocks,
  scenarios,
  alternatives,
  attribute,
  attribute_short
)
}
\arguments{
\item{csv_file}{The file we want to process}

\item{total_scenarios}{A list containing all the Stated Preference scenarios}

\item{num_blocks}{The amount of blocks on which the scenarios will be distributed}

\item{scenarios}{A list containing the relevant Stated Preference scenarios for the specific attribute of Lexicographic Response Styles}

\item{alternatives}{A list containing the alternatives for the Stated Preference experiment}

\item{attribute}{The attribute chosen for Lexicographic Response Styles}

\item{attribute_short}{The short form of the attribute name (for table names)
lrs_general_function()}
}
\description{
This function detects the responses employing lexicographic response styles, where the respondents choose the option for stated preference
scenarios based on the best attribute level.
}
\examples{
\dontrun{
lrs_general_function(file.csv, list(list("S1", "S2"),
                                    list("S3", "S4"),
                                    list("S5", "S6"),
                                    list("S7", "S8")),
                                    4,
                                    list(list("S1"),
                                         list("S4"),
                                         list("S6"),
                                         list("S7")),
                                         "Cheapest_Cost",
                                         "CC")
lrs_general_function(file.csv, list(list("S1", "S2", "S3", "S4"),
                                    list("S5", "S6", "S7", "S8")),
                                    2,
                                    list(list("S1", "S3"),
                                         list("S7", "S8")),
                                         "Fastest_Travel_Time",
                                         "FTT")
}
}
