% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasdetection_functions.R
\name{mrs_function}
\alias{mrs_function}
\title{Mid point Response Styles}
\usage{
mrs_function(csv_file, likert_columns, mid_value)
}
\arguments{
\item{csv_file}{The file we want to process}

\item{likert_columns}{A list containing the column names to use for MRS}

\item{mid_value}{Mid rate in the likert scale}
}
\value{
A list containing the updated data
mrs_function()
}
\description{
Mid point Response Styles
}
