% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasdetection_functions.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove duplicate}
\usage{
remove_duplicates(csv_data)
}
\arguments{
\item{csv_data}{A list containing the responses}
}
\value{
A list containing unique responses (based on id) from the original panel data
remove_duplicates()
}
\description{
Removes the duplicate entries from the panel data based on the id of the respondent to detect random responses.
}
