% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasdetection_functions.R
\name{rr_time_filter}
\alias{rr_time_filter}
\title{Random response time filter}
\usage{
rr_time_filter(rr_data)
}
\arguments{
\item{rr_data}{A list containing unique responses (based on id) from the original panel data

rr_time_filter()}
}
\description{
Calculates the first quartiles of each Time Group to identify the random responses. It displays them as well.
}
